fix windows compile issure

+++ src/tsdb/src/tsdbCommit.c
@@ -17,9 +17,9 @@
 #define TSDB_MAX_SUBBLOCKS 8
 static FORCE_INLINE int TSDB_KEY_FID(TSKEY key, int32_t days, int8_t precision) {
   if (key < 0) {
-    return -((-key) / tsMsPerDay[precision] / days + 1);
+    return (int)(-((-key) / tsMsPerDay[precision] / days + 1));
   } else {
-    return (key / tsMsPerDay[precision] / days);
+    return (int)((key / tsMsPerDay[precision] / days));
   }
 }
